All working, with Modal and validation. Still have unknown error in console that does not seem to affect anything.

+++ public/index.js
@@ -1,46 +1,68 @@
 // Get the modal
 var modal = document.getElementById('myModal')
+var newContact = document.getElementById('new-contact-form')
+var details = document.getElementById('details-modal')
 // Get the button that opens the modal
-var modalBtn = document.getElementById('modalBtn');
+var modalNew = document.getElementById('modalNew');
 // Get the <span> element that closes the modal
 var closeBtn = document.getElementsByClassName('modal-close')[0];
 // Listen for open click
-modalBtn.addEventListener('click', openModal)
+modalNew.addEventListener('click', openNewContact)
 // Listen for close click
 closeBtn.addEventListener('click', closeModal)
 // Listen for outside click
 window.addEventListener('click', outsideClick)
+
 //function to open modal
-function openModal(){
+//open new contact open
+function openNewContact(){
     modal.style.display = 'block'
+    newContact.style.display ='block'
+    document.getElementById('index').style.filter = "blur(5px)"
+}
+//open details modal
+function openDetails(){
+    var detailName = event.target.closest('tr').getElementsByClassName('name')[0].innerHTML
+    var detailEmail = event.target.closest('tr').getElementsByClassName('email')[0].innerHTML
+    var detailPhone = event.target.closest('tr').getElementsByClassName('phone')[0].innerHTML
+    document.getElementById('detailName').innerHTML = detailName
+    document.getElementById('detailEmail').innerHTML = detailEmail
+    document.getElementById('detailPhone').innerHTML = detailPhone
+    modal.style.display = 'block'
+    details.style.display ='block'
     document.getElementById('index').style.filter = "blur(5px)"
 }
 // function to close modal
 function closeModal(){
     modal.style.display = 'none'
+    newContact.style.display = 'none'
+    details.style.display = 'none'
     document.getElementById('index').style.filter = "blur(0px)"
 }
 //function to close modal if outside click
 function outsideClick(e){
     if(e.target == modal){
         modal.style.display = 'none'
+        newContact.style.display = 'none'
+        details.style.display = 'none'
         document.getElementById('index').style.filter = "blur(0px)"
     }
 }
 $(document).ready(function(){
     
         $('.delete').on('click', function(){
-            var id = $(this).closest('tr').attr('id')
-            var phone = $(this).find('.phone').text()
-            //var item = ''
-            $.ajax({
-                type: 'DELETE',
-                url: '/'+id,
-                //data:phone,
-                success: function(data){
-                    location.reload()
-                }
-            })
+            var r = confirm('Are you sure you want to delete?')
+            if (r==true){
+                var id = $(this).closest('tr').attr('id')
+                var phone = $(this).find('.phone').text()
+                $.ajax({
+                    type: 'DELETE',
+                    url: '/'+id,
+                    success: function(data){
+                        location.reload()
+                    }
+                })
+            }
         })
     
         $('.edit').on('click', function(){
@@ -59,5 +81,4 @@ $(document).ready(function(){
                 }
             })
         })
-    
-    })
+})
